+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `keepachangelog.release` function to guess new version number based on `Unreleased` section, update changelog and return new version number.
 
 ## [0.4.0] - 2020-09-21
 ### Added